Added union types into the spec
The goal of this proposal is to be as forwards compatible as possible. Having a null-specific token (?) doesn't bode well with that idea. It just shows up in a lot of grammar rules. Also I need to get TypeScript users onboard with this and their language designers probably know best.